Further NSInteger-ization

+++ AppKit/NSAttributedString.m
@@ -137,7 +137,7 @@ NSString * const NSSpellingStateAttributeName=@"NSSpellingStateAttributeName"; /
 #pragma mark -
 #pragma mark Calculating Linguistic Units
 
--(NSRange)doubleClickAtIndex:(unsigned)location {
+-(NSRange)doubleClickAtIndex:(NSUInteger)location {
 	NSRange   result=NSMakeRange(location,0);
 	NSString *string=[self string];
 	unsigned  length=[string length];
@@ -171,12 +171,12 @@ NSString * const NSSpellingStateAttributeName=@"NSSpellingStateAttributeName"; /
 	return result;
 }
 
--(unsigned)lineBreakBeforeIndex:(unsigned)index withinRange:(NSRange)range {
+-(NSUInteger)lineBreakBeforeIndex:(NSUInteger)index withinRange:(NSRange)range {
 	NSUnimplementedMethod();
 	return 0;
 }
 
--(unsigned)lineBreakByHyphenatingBeforeIndex:(unsigned)index withinRange:(NSRange)range {
+-(NSUInteger)lineBreakByHyphenatingBeforeIndex:(NSUInteger)index withinRange:(NSRange)range {
 	NSUnimplementedMethod();
 	return 0;
 }
@@ -258,22 +258,22 @@ NSString * const NSSpellingStateAttributeName=@"NSSpellingStateAttributeName"; /
 #pragma mark -
 #pragma mark Calculating Ranges
 
--(int)itemNumberInTextList:(NSTextList *)list atIndex:(unsigned)index {
+-(NSInteger)itemNumberInTextList:(NSTextList *)list atIndex:(NSUInteger)index {
    NSUnimplementedMethod();
    return 0;
 }
 
--(NSRange)rangeOfTextBlock:(NSTextBlock *)block atIndex:(unsigned)index {
+-(NSRange)rangeOfTextBlock:(NSTextBlock *)block atIndex:(NSUInteger)index {
    NSUnimplementedMethod();
    return NSMakeRange(0,0);
 }
 
--(NSRange)rangeOfTextList:(NSTextList *)list atIndex:(unsigned)index {
+-(NSRange)rangeOfTextList:(NSTextList *)list atIndex:(NSUInteger)index {
    NSUnimplementedMethod();
    return NSMakeRange(0,0);
 }
 
--(NSRange)rangeOfTextTable:(NSTextTable *)table atIndex:(unsigned)index {
+-(NSRange)rangeOfTextTable:(NSTextTable *)table atIndex:(NSUInteger)index {
    NSUnimplementedMethod();
    return NSMakeRange(0,0);
 }
